ml_service/requirements.txt
tensorflow
kafka-python
numpy
joblib
scikit-learn
pandas